[MOVE/#75] 투두 조회뷰 재분배
